Move i.MX ethernet driver include out of util directory
